Panic if JS git hash doesn’t match Rust git hash
Closes #26.

This is to prevent accidental mistakes where the two parts of Zaplib are
out of sync.

For local development to not suck, we also have to restart the webpack
server any time the git hash changes. I tried to keep that logic within
Webpack itself, but none of the several packages I tried worked.. I
ended up wrapping the Webpack command itself in “nodemon”, which has the
added benefit that we can also restart Webpack when webpack.config.js
changes.

I also extracted out some common logic in webpack.config.